Fixed compile error

+++ src/test/scala/firepile/tests/TestTimeMap2c.scala
@@ -43,8 +43,8 @@ object TestTimeMap2c {
           k
       }
 
-      val bx: BBArray[Float] = a
-      val b: BBArray[Float] = bx.directCopy
+      val bx: BBArray[Int] = a
+      val b: BBArray[Int] = bx.directCopy
 
       println(n + " bbarray map x2");
       time {
